requirements.txt
pandas==2.1.4
streamlit==1.29.0
altair==5.2.0
Pillow==10.1.0
